moved plot functions to outside Debugger class reordered the mess in map_data init,  moved set pathlib to Pather
